Initial commit
This commit sets up the basic structure of the repository, as well as
breaking the project down into a common library, commandline program for
servers wanting to run a distrib node, and a set of example programs
showing how the library is intended to be used. Documentation and
testing directories are also created, and a set of build scripts are
written for building the project and packaging it into an archive.